refactor(notion-fetch): extract shared markdown code-block masking utility
Both contentSanitizer and linkNormalizer implemented nearly identical
logic for masking and restoring fenced code blocks and inline code spans.
Extract into a shared markdownUtils.ts to eliminate duplication and
prevent future drift.
