+++ scripts/notion-fetch/contentSanitizer.ts
@@ -4,6 +4,11 @@
  */
 
 import { createSafeSlug } from "./slugUtils";
+import {
+  maskFencedCodeBlocks,
+  maskInlineCodeSpans,
+  restoreCodeMasks,
+} from "./markdownUtils";
 
 const EMOJI_STYLE_MARKERS = ["display:", "height:", "margin:"];
 
@@ -70,184 +75,6 @@ function fixHeadingHierarchy(
   return fixedLines.join("\n");
 }
 
-function createCodeBlockMasker(content: string): {
-  content: string;
-  codeBlocks: string[];
-  codeBlockPlaceholders: string[];
-} {
-  const lines = content.split("\n");
-  const codeBlocks: string[] = [];
-  const codeBlockPlaceholders: string[] = [];
-  const maskedLines: string[] = [];
-
-  let inFence = false;
-  let fenceChar = "";
-  let fenceLength = 0;
-  let blockLines: string[] = [];
-
-  for (const line of lines) {
-    if (!inFence) {
-      const openingMatch = line.match(/^ {0,3}(`{3,}|~{3,})(.*)$/);
-
-      if (!openingMatch) {
-        maskedLines.push(line);
-        continue;
-      }
-
-      const fence = openingMatch[1];
-      fenceChar = fence[0];
-      fenceLength = fence.length;
-      blockLines = [line];
-      inFence = true;
-      continue;
-    }
-
-    blockLines.push(line);
-
-    if (isClosingFenceLine(line, fenceChar, fenceLength)) {
-      codeBlocks.push(blockLines.join("\n"));
-      const placeholder = `__CODEBLOCK_${codeBlocks.length - 1}__`;
-      codeBlockPlaceholders.push(placeholder);
-      maskedLines.push(placeholder);
-      inFence = false;
-      blockLines = [];
-    }
-  }
-
-  if (inFence) {
-    codeBlocks.push(blockLines.join("\n"));
-    const placeholder = `__CODEBLOCK_${codeBlocks.length - 1}__`;
-    codeBlockPlaceholders.push(placeholder);
-    maskedLines.push(placeholder);
-  }
-
-  return {
-    content: maskedLines.join("\n"),
-    codeBlocks,
-    codeBlockPlaceholders,
-  };
-}
-
-function isClosingFenceLine(
-  line: string,
-  fenceChar: string,
-  fenceLength: number
-): boolean {
-  let i = 0;
-
-  while (i < line.length && line.charAt(i) === " ") {
-    i++;
-  }
-
-  if (i > 3) {
-    return false;
-  }
-
-  let fenceCount = 0;
-  while (i < line.length && line.charAt(i) === fenceChar) {
-    fenceCount++;
-    i++;
-  }
-
-  if (fenceCount < fenceLength) {
-    return false;
-  }
-
-  while (i < line.length) {
-    if (line.charAt(i) !== " " && line.charAt(i) !== "\t") {
-      return false;
-    }
-    i++;
-  }
-
-  return true;
-}
-
-function maskCodeFences(content: string): {
-  content: string;
-  codeBlocks: string[];
-  codeBlockPlaceholders: string[];
-} {
-  return createCodeBlockMasker(content);
-}
-
-function restoreCodeFences(content: string, codeBlocks: string[]): string {
-  let restoredContent = content;
-  for (const [index, codeBlock] of codeBlocks.entries()) {
-    restoredContent = restoredContent.replaceAll(
-      `__CODEBLOCK_${index}__`,
-      codeBlock
-    );
-  }
-  return restoredContent;
-}
-
-function maskInlineCodeSpans(content: string): {
-  content: string;
-  codeSpans: string[];
-} {
-  const codeSpans: string[] = [];
-  const output: string[] = [];
-
-  let i = 0;
-  while (i < content.length) {
-    const currentChar = content.charAt(i);
-    if (currentChar !== "`") {
-      output.push(currentChar);
-      i++;
-      continue;
-    }
-
-    let openingLength = 0;
-    while (
-      i + openingLength < content.length &&
-      content.charAt(i + openingLength) === "`"
-    ) {
-      openingLength++;
-    }
-
-    let scanIndex = i + openingLength;
-    let closingIndex = -1;
-    while (scanIndex < content.length) {
-      const nextBacktick = content.indexOf("`", scanIndex);
-      if (nextBacktick === -1) {
-        break;
-      }
-
-      let closingLength = 0;
-      while (
-        nextBacktick + closingLength < content.length &&
-        content.charAt(nextBacktick + closingLength) === "`"
-      ) {
-        closingLength++;
-      }
-
-      if (closingLength === openingLength) {
-        closingIndex = nextBacktick;
-        break;
-      }
-
-      scanIndex = nextBacktick + closingLength;
-    }
-
-    if (closingIndex === -1) {
-      output.push(content.slice(i, i + openingLength));
-      i += openingLength;
-      continue;
-    }
-
-    const codeSpan = content.slice(i, closingIndex + openingLength);
-    codeSpans.push(codeSpan);
-    output.push(`__CODESPAN_${codeSpans.length - 1}__`);
-    i = closingIndex + openingLength;
-  }
-
-  return {
-    content: output.join(""),
-    codeSpans,
-  };
-}
-
 export function injectExplicitHeadingIds(content: string): string {
   if (!content) {
     return content;
@@ -256,16 +83,14 @@ export function injectExplicitHeadingIds(content: string): string {
   const {
     content: maskedContent,
     codeBlocks,
-    codeBlockPlaceholders,
-  } = maskCodeFences(content);
+    placeholders,
+  } = maskFencedCodeBlocks(content);
   const reservedIds = new Set<string>();
   const headingCounts = new Map<string, number>();
 
   const lines = maskedContent.split("\n");
   for (const line of lines) {
-    if (
-      codeBlockPlaceholders.some((placeholder) => line.includes(placeholder))
-    ) {
+    if (placeholders.some((placeholder) => line.includes(placeholder))) {
       continue;
     }
 
@@ -288,9 +113,7 @@ export function injectExplicitHeadingIds(content: string): string {
   }
 
   const updatedLines = lines.map((line) => {
-    if (
-      codeBlockPlaceholders.some((placeholder) => line.includes(placeholder))
-    ) {
+    if (placeholders.some((placeholder) => line.includes(placeholder))) {
       return line;
     }
 
@@ -326,7 +149,7 @@ export function injectExplicitHeadingIds(content: string): string {
     return `${leadingWhitespace}${hashes} ${headingText} {#${headingId}}`;
   });
 
-  return restoreCodeFences(updatedLines.join("\n"), codeBlocks);
+  return restoreCodeMasks(updatedLines.join("\n"), codeBlocks, []);
 }
 
 /**
@@ -341,14 +164,14 @@ export function sanitizeMarkdownContent(content: string): string {
   const {
     content: maskedContent,
     codeBlocks,
-    codeBlockPlaceholders,
-  } = maskCodeFences(content);
+    placeholders,
+  } = maskFencedCodeBlocks(content);
   const { content: maskedWithCodeSpans, codeSpans } =
     maskInlineCodeSpans(maskedContent);
   content = maskedWithCodeSpans;
 
   // 1. Fix heading hierarchy for proper TOC generation (after masking code blocks)
-  content = fixHeadingHierarchy(content, codeBlockPlaceholders);
+  content = fixHeadingHierarchy(content, placeholders);
 
   // 2. Aggressively strip all curly-brace expressions by unwrapping to inner text
   // BUT preserve JSX style objects for emoji images
@@ -412,15 +235,7 @@ export function sanitizeMarkdownContent(content: string): string {
   }
 
   // 9. Restore masked code blocks and inline code
-  content = restoreCodeFences(content, codeBlocks);
-  let restoredContent = content;
-  for (const [index, codeSpan] of codeSpans.entries()) {
-    restoredContent = restoredContent.replaceAll(
-      `__CODESPAN_${index}__`,
-      codeSpan
-    );
-  }
-  content = restoredContent;
+  content = restoreCodeMasks(content, codeBlocks, codeSpans);
 
   return content;
 }
