+++ scripts/notion-fetch/linkNormalizer.ts
@@ -1,5 +1,10 @@
 import config from "../../docusaurus.config";
 import { createSafeSlug } from "./slugUtils";
+import {
+  maskFencedCodeBlocks,
+  maskInlineCodeSpans,
+  restoreCodeMasks,
+} from "./markdownUtils";
 
 const DEFAULT_LOCALE = config.i18n.defaultLocale;
 const MARKDOWN_LINK_REGEX = /(?<![!])\[([^\]]+)\]\(([^)\n]+)\)/gm;
@@ -10,135 +15,6 @@ function safeDecode(s: string): string {
   } catch {
     return s;
   }
-}
-
-function maskFencedCodeBlocks(content: string): {
-  maskedContent: string;
-  codeBlocks: string[];
-} {
-  const codeBlocks: string[] = [];
-  const lines = content.split("\n");
-  const output: string[] = [];
-
-  let inFence = false;
-  let fenceChar = "";
-  let fenceLength = 0;
-  let fencedBlock: string[] = [];
-
-  for (const line of lines) {
-    if (!inFence) {
-      const openMatch = /^ {0,3}(`{3,}|~{3,})(.*)$/.exec(line);
-      if (openMatch) {
-        inFence = true;
-        fenceChar = openMatch[1][0];
-        fenceLength = openMatch[1].length;
-        fencedBlock = [line];
-        continue;
-      }
-
-      output.push(line);
-      continue;
-    }
-
-    fencedBlock.push(line);
-
-    const closeMatch = /^ {0,3}([`~]{3,})\s*$/.exec(line);
-    if (
-      closeMatch &&
-      closeMatch[1][0] === fenceChar &&
-      closeMatch[1].length >= fenceLength
-    ) {
-      codeBlocks.push(fencedBlock.join("\n"));
-      output.push(`__LINK_NORMALIZER_CODEBLOCK_${codeBlocks.length - 1}__`);
-      inFence = false;
-      fenceChar = "";
-      fenceLength = 0;
-      fencedBlock = [];
-    }
-  }
-
-  if (inFence) {
-    codeBlocks.push(fencedBlock.join("\n"));
-    output.push(`__LINK_NORMALIZER_CODEBLOCK_${codeBlocks.length - 1}__`);
-  }
-
-  return { maskedContent: output.join("\n"), codeBlocks };
-}
-
-function maskInlineCode(content: string): {
-  maskedContent: string;
-  codeSpans: string[];
-} {
-  const codeSpans: string[] = [];
-  const output: string[] = [];
-
-  let index = 0;
-
-  while (index < content.length) {
-    const char = content.charAt(index);
-    if (char !== "`") {
-      output.push(char);
-      index++;
-      continue;
-    }
-
-    let openerLength = 1;
-    while (content.charAt(index + openerLength) === "`") {
-      openerLength++;
-    }
-
-    let cursor = index + openerLength;
-    let closingIndex = -1;
-    while (cursor < content.length) {
-      if (content.charAt(cursor) !== "`") {
-        cursor++;
-        continue;
-      }
-
-      let runLength = 1;
-      while (content.charAt(cursor + runLength) === "`") {
-        runLength++;
-      }
-
-      if (runLength === openerLength) {
-        closingIndex = cursor;
-        break;
-      }
-
-      cursor += runLength;
-    }
-
-    if (closingIndex === -1) {
-      output.push(content.slice(index));
-      break;
-    }
-
-    const codeSpan = content.slice(index, closingIndex + openerLength);
-    codeSpans.push(codeSpan);
-    output.push(`__LINK_NORMALIZER_CODESPAN_${codeSpans.length - 1}__`);
-    index = closingIndex + openerLength;
-  }
-
-  return { maskedContent: output.join(""), codeSpans };
-}
-
-function restoreCode(
-  content: string,
-  codeBlocks: string[],
-  codeSpans: string[]
-): string {
-  const restoreByIndex = (values: string[], rawIndex: string) => {
-    const index = Number(rawIndex);
-    return Number.isInteger(index) ? (values.at(index) ?? "") : "";
-  };
-
-  return content
-    .replace(/__LINK_NORMALIZER_CODESPAN_(\d+)__/g, (_match, index) => {
-      return restoreByIndex(codeSpans, index);
-    })
-    .replace(/__LINK_NORMALIZER_CODEBLOCK_(\d+)__/g, (_match, index) => {
-      return restoreByIndex(codeBlocks, index);
-    });
 }
 
 function normalizeDocPathname(pathname: string): string {
@@ -180,9 +56,9 @@ export function normalizeInternalDocLinks(
     return content;
   }
 
-  const { maskedContent: maskedBlocks, codeBlocks } =
-    maskFencedCodeBlocks(content);
-  const { maskedContent, codeSpans } = maskInlineCode(maskedBlocks);
+  const { content: maskedBlocks, codeBlocks } = maskFencedCodeBlocks(content);
+  const { content: maskedContent, codeSpans } =
+    maskInlineCodeSpans(maskedBlocks);
 
   const normalizedContent = maskedContent.replace(
     MARKDOWN_LINK_REGEX,
@@ -204,5 +80,5 @@ export function normalizeInternalDocLinks(
     }
   );
 
-  return restoreCode(normalizedContent, codeBlocks, codeSpans);
+  return restoreCodeMasks(normalizedContent, codeBlocks, codeSpans);
 }
